Fixed a bug where the popular widget was wrong
Turned out the last story listed in opinion would sometimes override the
settings of the popular posts widget. This is because each story on the
front page has its own query without resetting the global WP query. Now
the WP query is reset after all custom things are done so the popular
widget can rely on having a clean query.